More minor polishes
- Adjusted Eye and Hair sprites
- LoadSpriteAnimations can takeWidth/Height as params
- Increased starting Energy
- Energy Bar only shows int
- Reduced Scythe energy cost
- No energy lost on a miss
- Exit Build Mode on go To Title
- Can cancel out of building a Building

+++ RiverHollowGame/GUIComponents/GUIObjects/GUIWindows/GUIOldStatDisplay.cs
@@ -58,7 +58,7 @@ namespace RiverHollow.GUIComponents.GUIObjects.GUIWindows
 
         protected override void BeginHover()
         {
-            _gText.SetText(string.Format("{0}/{1}", _fCurr, _fMax));
+            _gText.SetText(string.Format("{0}/{1}", (int)_fCurr, (int)_fMax));
             _gText.AlignToObject(this, SideEnum.Center);
             _gText.Show(true);
         }
